enhanced recommendation system. Now its based on views/likes/participations and is faster coz of using Locality-sensitive hashing

+++ src/main/java/com/hse/DAOs/RecommendationDao.java
@@ -18,7 +18,7 @@ public class RecommendationDao {
         namedJdbcTemplate.update("TRUNCATE users_with_recommendations", map);
     }
 
-    public void addEventRecommendation(Long userId, Long eventId, float coefficient) {
+    public void addEventRecommendation(Long userId, Long eventId, double coefficient) {
         MapSqlParameterSource map = new MapSqlParameterSource();
         map.addValue("userId", userId);
         map.addValue("eventId", eventId);
